refactor: Added legacy support option `fixNextedATags`
